[App] 引入 TableLayout, 简化 Log 界面布局; 修复默认主题名称错误导致无法找到主题; ThemeManager 各注册监听方法新增对应的取消注册方法; 移除 LogCmdManager 相关控件和代码; 移除 retryAdb 逻辑; 移除 GetDeviceTask, 改为由 DeviceManager 实现; 新增 Script 相关界面和类; 修复 DeviceCombo 选中状态边宽显示不全问题 (editor 要使用继承自 UIResource 的) [Utils] Context 框架, 移除 ContextScope 类;

+++ utils/src/main/kotlin/me/gegenbauer/catspy/context/ContextManager.kt
@@ -2,7 +2,7 @@ package me.gegenbauer.catspy.context
 
 interface ContextManager {
 
-    fun getContext(id: Int): Context?
+    fun getContext(id: Long): Context?
 
     fun register(context: Context)
